Move all panels to a panels folder, delete unused panels
