+++ sdk/react/panels/style/discoveryPanelStyles.json
