+++ sdk/react/panels/style/endScreenStyles.json
