+++ sdk/react/panels/style/errorScreenStyles.json
