+++ sdk/react/panels/style/moreOptionScreenStyles.json
