+++ sdk/react/panels/style/panelStyles.json
