+++ sdk/react/panels/style/startScreenStyles.json
