+++ sdk/react/panels/style/videoViewStyles.json
